ipq40xx: Mikrotik Wireless Wire Dish LHGG-60ad: Fix images not booting
This patch fixes the issue with aux-loader that images will not boot
if UART has not been enabled on the device as Mikrotik ships them will UART
disabled.
This makes the aux-loader fail writing to UART and watchdog will kick in resetting
the device.
So lets pass flag that will disable UART in aux-loader, but in OpenWrt it will still work.

Signed-off-by: Robert Marko <robimarko@gmail.com>